requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.0
pydantic>=1.10.0
httpx>=0.27.0
pydantic-settings
SQLAlchemy>=2.0.0
PyJWT>=2.9.0
